# docs/swagger.yaml
basePath: /v1
definitions:
...
  entity.Category:
    properties:
      created_at:
        type: string
      id:
        type: string
      name:
        type: string
      updated_at:
        type: string
    type: object
  entity.CategoryList:
    properties:
      count:
        type: integer
      items:
        items:
          $ref: '#/definitions/entity.Category'
        type: array
    type: object
  entity.ErrorResponse:
    properties:
      code:
        type: string
      message:
        type: string
    type: object
...
  entity.Notification:
    properties:
      created_at:
        type: string
      email:
        type: string
      id:
        type: string
      message:
        type: string
      owner_id:
        type: string
      ownerrole:
        type: string
      status:
        type: string
      user_id:
        type: string
    type: object
  entity.NotificationList:
    properties:
      count:
        type: integer
      notifications:
        items:
          $ref: '#/definitions/entity.Notification'
        type: array
    type: object
...
  entity.SuccessResponse:
    properties:
      message:
        type: string
    type: object
...
info:
  contact: {}
  description: This is a sample server Food Delivery server.
  title: Food Delivery API
  version: "1.0"
paths:
...
  /category:
    post:
      consumes:
      - application/json
      description: Create a new category
      parameters:
      - description: Category object
        in: body
        name: category
        required: true
        schema:
          $ref: '#/definitions/entity.Category'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/entity.Category'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/entity.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create a new category
      tags:
      - category
    put:
      consumes:
      - application/json
      description: Update a category
      parameters:
      - description: Category object
        in: body
        name: category
        required: true
        schema:
          $ref: '#/definitions/entity.Category'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.Category'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/entity.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update a category
      tags:
      - category
  /category/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a category
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/entity.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete a category
      tags:
      - category
    get:
      consumes:
      - application/json
      description: Get a category by ID
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.Category'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/entity.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get a category by ID
      tags:
      - category
  /category/list:
    get:
      consumes:
      - application/json
      description: Get a list of categorys
      parameters:
      - description: page
        in: query
        name: page
        required: true
        type: number
      - description: limit
        in: query
        name: limit
        required: true
        type: number
      - description: search
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.CategoryList'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/entity.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get a list of categorys
      tags:
      - category
...
  /notification:
    post:
      consumes:
      - application/json
      description: Create a notification and send an email to user
      parameters:
      - description: Notification data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/entity.Notification'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/entity.Notification'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/entity.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create a new notification
      tags:
      - notification
    put:
      consumes:
      - application/json
      description: Update a notification
      parameters:
      - description: Notification object
        in: body
        name: notification
        required: true
        schema:
          $ref: '#/definitions/entity.Notification'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.Notification'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/entity.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update a notification
      tags:
      - notification
  /notification/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a notification
      parameters:
      - description: Notification ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/entity.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete a notification
      tags:
      - notification
    get:
      consumes:
      - application/json
      description: Get a notification by ID
      parameters:
      - description: Notification ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.Notification'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/entity.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get a notification by ID
      tags:
      - notification
  /notification/list:
    get:
      consumes:
      - application/json
      description: Get a list of notifications
      parameters:
      - description: page
        in: query
        name: page
        required: true
        type: number
      - description: limit
        in: query
        name: limit
        required: true
        type: number
      - description: user_id
        in: query
        name: user_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.NotificationList'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/entity.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get a list of notifications
      tags:
      - notification
  /notification/status:
    put:
      consumes:
      - application/json
      description: Update a notification's status
      parameters:
      - description: Notification object
        in: body
        name: notification
        required: true
        schema:
          $ref: '#/definitions/entity.Notification'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.Notification'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/entity.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update a notification status
      tags:
      - notification
  /notification/update-status:
    put:
      consumes:
      - application/json
      description: Update status notification by ID
      parameters:
      - description: Notification object
        in: body
        name: notification
        required: true
        schema:
          $ref: '#/definitions/entity.Notification'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.Notification'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/entity.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update status a notification by ID
      tags:
      - notification
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"